% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointMR_analysis.R
\name{JointMR_analysis}
\alias{JointMR_analysis}
\title{Estimate the causal effect using JointMR}
\usage{
JointMR_analysis(
  exposure_filter_list,
  outcome_filter_list,
  original_outcome_list,
  N_list,
  ancestry = "EUR",
  bootstrap_time = 500,
  relaxed_nome_bootstrap_time = NULL,
  original_exposure_list = NULL,
  ldsc_cache = NULL,
  run_relaxed_nome = TRUE,
  ldsc_scope = c("all", "yy")
)
}
\arguments{
\item{exposure_filter_list}{A list of filtered exposure GWAS data frames.}

\item{outcome_filter_list}{A list of filtered outcome GWAS data frames.}

\item{original_outcome_list}{A list of original outcome GWAS data frames.}

\item{N_list}{A list containing sample sizes for LDSC estimation.}

\item{ancestry}{Character, e.g., "EUR".}

\item{bootstrap_time}{Integer, number of bootstrap iterations.}

\item{relaxed_nome_bootstrap_time}{Integer, number of relaxed NOME bootstraps.}

\item{original_exposure_list}{Optional list of original exposure GWAS data.}

\item{ldsc_cache}{Optional cached LDSC correlation matrices.}

\item{run_relaxed_nome}{Logical, whether to run the relaxed-NOME extension.}

\item{ldsc_scope}{Character, scope for LDSC correlation estimation ("all" or "yy").}
}
\description{
This function integrates repeated MR information obtained from all
exposure-outcome GWAS database pairings.
}
