% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_exposure}
\alias{sample_exposure}
\title{Sample exposure GWAS datasets}
\format{
A list of data frames.
}
\source{
Simulated data for demonstration purposes.
}
\usage{
sample_exposure
}
\description{
A simulated list of exposure GWAS datasets for the JointMR tutorial.
}
\keyword{datasets}
