% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_N_list}
\alias{sample_N_list}
\title{Sample sizes for the mock GWAS datasets}
\format{
A list with numeric values.
}
\source{
Simulated data for demonstration purposes.
}
\usage{
sample_N_list
}
\description{
A list of sample sizes corresponding to the simulated exposure and outcome datasets.
}
\keyword{datasets}
