% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_outcome}
\alias{sample_outcome}
\title{Sample outcome GWAS datasets}
\format{
A list of data frames.
}
\source{
Simulated data for demonstration purposes.
}
\usage{
sample_outcome
}
\description{
A simulated list of outcome GWAS datasets for the JointMR tutorial.
}
\keyword{datasets}
